Delay containers to components directory refactor
