+++ src/containers/SinglePost/connector.js
